Minimize APPWIDGET_UPDATE in ordered broadcast queue
Widget providers would receive all APPWIDGET_UPDATE intent
which is not set with a target component name by the sender.
This is a waste of the system resource, and resulting the
poor performance.

Solution:
Before broadcast the intent APPWIDGET_UPDATE, minimize the
receivers in the list by removing those ones the widget ID was
not part of the target widget specified in the intent extra fields.

Change-Id: I3c772062e039aabeb257ccc063f6b0d88ca51bb5