QuickSettings: Ram Tile WIP Dont Merge 1/3
Need to do.

Clean up layout/ make moar prettier
Come up with a way to automagiclly turn the listener on and off

For now you can turn it on and off with a click

Ram tile and Ram bar can not run at the same time. one will
not update if the other is running.

Change-Id: I6940af075aa5a856cfd0056460e7395967d4a3bb
Signed-off-by: Stevespear426 <stephen.k.spear@gmail.com>

Conflicts:
	packages/SystemUI/src/com/android/systemui/statusbar/phone/QuickSettings.java
	packages/SystemUI/src/com/android/systemui/statusbar/phone/QuickSettingsModel.java

Conflicts:
	packages/SystemUI/src/com/android/systemui/statusbar/phone/QuickSettingsModel.java